Moved custom buttons/text to Styles folder. Improved custom button styling.
